+++ src/imio/helpers/tests/test_security.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import os
+import re
 from AccessControl.SecurityManagement import getSecurityManager
 
 from imio.helpers.testing import IntegrationTestCase
@@ -106,3 +107,12 @@ class TestSecurityModule(IntegrationTestCase):
         self.assertFalse(is_develop_environment())
         os.environ['IS_DEV_ENV'] = 'true'
         self.assertTrue(is_develop_environment())
+
+    def test_generate_password(self):
+        from imio.helpers.security import generate_password
+        pwd = generate_password(length=10, digits=1, upper=1, lower=1, special=1)
+        self.assertEqual(len(pwd), 10)
+        self.assertIsNotNone(re.search(r'\d', pwd))
+        self.assertIsNotNone(re.search(r'[A-Z]', pwd))
+        self.assertIsNotNone(re.search(r'[a-z]', pwd))
+        self.assertIsNotNone(re.search(r'[!#$%&*+<=>?@-]', pwd))